# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/postgallery": {
            "post": {
                "tags": [
                    "upload gallery images"
                ],
                "summary": "Upload gallery images",
                "description": "Endpoint for uploading images",
                "operationId": "storeGallery",
                "requestBody": {
                    "description": "Data for uploading images",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Image Description",
                                        "type": "string"
                                    },
                                    "picture": {
                                        "description": "Image Files",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}